requirements.txt
databricks-sdk>=0.4.0, <1.0.0
gitpython
pandas
solara>=1.19.0, <2.0.0
plotly
pyarrow
pydantic
configparser
sqlalchemy
tabulate